Implementing requested changes.
